requirements.txt
Flask
discord.py
selenium
python-dotenv